AI_Projects/neuro_video_copywriter/backend/pyproject.toml
[tool.poetry]
name = "neuro-video-copywriter-backend"
version = "0.1.0"
description = "Backend for Neuro Video Copywriter project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "backend" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
pydantic = "^2.6.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg = { version = "^3.1.18", extras = ["binary"] }
qdrant-client = "^1.7.3"
redis = "^5.0.3"
yt-dlp = "^2024.4.9"
openai-whisper = "^20231117"
openai = "^1.12.0"
pydub = "^0.25.1"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
ruff = "^0.3.3"
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

